requirements.txt
psycopg2-binary
python-dotenv
groq

pytest-cov